+++ ACD3100.cpp
@@ -181,32 +181,6 @@ uint8_t ACD3100::getRequestTime()
 //
 //  CALIBRATION
 //
-bool ACD3100::setCalibrationMode(uint8_t mode)
-{
-  if (mode > 1) return false;
-  uint8_t buf[5] = { 0x53, 0x06, 0x00, 0x00, 0x00 };
-  buf[3] = mode;
-  buf[4] = _crc8(&buf[2], 2);
-  //  Serial.println(buf[4], HEX);
-  _command(buf, 5);
-  return true;
-}
-
-
-uint8_t ACD3100::readCallibrationMode()
-{
-  uint8_t buf[3] = { 0x53, 0x06, 0x00 };
-  _command(buf, 2);
-  _request(buf, 3);
-  // if (buf[2] != _crc8(&buf[0], 2))
-  // {
-    // Serial.print(__FUNCTION__);
-    // Serial.println(": CRC error");
-  // }
-  return buf[1];
-}
-
-
 bool ACD3100::setManualCalibration(uint16_t value)
 {
   if ((value < 400) || (value > 5000)) return false;
@@ -255,7 +229,7 @@ bool ACD3100::readFactorySet()
     // Serial.print(__FUNCTION__);
     // Serial.println(": CRC error");
   // }
-  return (buf[1] == 0x01);
+  return ((buf[0] == 0) && (buf[1] == 0x01));
 }